Moved enums from baseclasses.py to enums.py and added more entries
